move bios and uefi specific code into submodules gated by features
