+++ build.rs
@@ -44,6 +44,7 @@ fn main() {
 }
 
 #[cfg(not(docsrs_dummy_build))]
+#[cfg(feature = "uefi")]
 fn build_uefi_bootloader(out_dir: &Path) -> PathBuf {
     let cargo = std::env::var("CARGO").unwrap_or_else(|_| "cargo".into());
     let mut cmd = Command::new(cargo);
@@ -78,6 +79,7 @@ fn build_uefi_bootloader(out_dir: &Path) -> PathBuf {
 }
 
 #[cfg(not(docsrs_dummy_build))]
+#[cfg(feature = "bios")]
 fn build_bios_boot_sector(out_dir: &Path) -> PathBuf {
     let cargo = std::env::var("CARGO").unwrap_or_else(|_| "cargo".into());
     let mut cmd = Command::new(cargo);
@@ -121,6 +123,7 @@ fn build_bios_boot_sector(out_dir: &Path) -> PathBuf {
 }
 
 #[cfg(not(docsrs_dummy_build))]
+#[cfg(feature = "bios")]
 fn build_bios_stage_2(out_dir: &Path) -> PathBuf {
     let cargo = std::env::var("CARGO").unwrap_or_else(|_| "cargo".into());
     let mut cmd = Command::new(cargo);
@@ -162,6 +165,7 @@ fn build_bios_stage_2(out_dir: &Path) -> PathBuf {
 }
 
 #[cfg(not(docsrs_dummy_build))]
+#[cfg(feature = "bios")]
 fn build_bios_stage_3(out_dir: &Path) -> PathBuf {
     let cargo = std::env::var("CARGO").unwrap_or_else(|_| "cargo".into());
     let mut cmd = Command::new(cargo);
@@ -203,6 +207,7 @@ fn build_bios_stage_3(out_dir: &Path) -> PathBuf {
 }
 
 #[cfg(not(docsrs_dummy_build))]
+#[cfg(feature = "bios")]
 fn build_bios_stage_4(out_dir: &Path) -> PathBuf {
     let cargo = std::env::var("CARGO").unwrap_or_else(|_| "cargo".into());
     let mut cmd = Command::new(cargo);
@@ -244,6 +249,7 @@ fn build_bios_stage_4(out_dir: &Path) -> PathBuf {
     convert_elf_to_bin(elf_path)
 }
 
+#[cfg(feature = "bios")]
 fn convert_elf_to_bin(elf_path: PathBuf) -> PathBuf {
     let flat_binary_path = elf_path.with_extension("bin");
 
@@ -273,23 +279,23 @@ fn convert_elf_to_bin(elf_path: PathBuf) -> PathBuf {
 
 // dummy implementations because docsrs builds have no network access
 
-#[cfg(docsrs_dummy_build)]
+#[cfg(any(docsrs_dummy_build, not(feature = "uefi")))]
 fn build_uefi_bootloader(_out_dir: &Path) -> PathBuf {
     PathBuf::new()
 }
-#[cfg(docsrs_dummy_build)]
+#[cfg(any(docsrs_dummy_build, not(feature = "bios")))]
 fn build_bios_boot_sector(_out_dir: &Path) -> PathBuf {
     PathBuf::new()
 }
-#[cfg(docsrs_dummy_build)]
+#[cfg(any(docsrs_dummy_build, not(feature = "bios")))]
 fn build_bios_stage_2(_out_dir: &Path) -> PathBuf {
     PathBuf::new()
 }
-#[cfg(docsrs_dummy_build)]
+#[cfg(any(docsrs_dummy_build, not(feature = "bios")))]
 fn build_bios_stage_3(_out_dir: &Path) -> PathBuf {
     PathBuf::new()
 }
-#[cfg(docsrs_dummy_build)]
+#[cfg(any(docsrs_dummy_build, not(feature = "bios")))]
 fn build_bios_stage_4(_out_dir: &Path) -> PathBuf {
     PathBuf::new()
 }
